+++ src/graph.rs
@@ -1022,16 +1022,8 @@ impl DirectedGraphTrait for DirectedGraph2 {
         }
 
         // assume we never get inconsistant
-        self.edges
-            .entry(vertex2)
-            .or_default()
-            .0
-            .push(vertex1);
-        self.edges
-            .entry(vertex2)
-            .or_default()
-            .0
-            .sort();
+        self.edges.entry(vertex2).or_default().0.push(vertex1);
+        self.edges.entry(vertex2).or_default().0.sort();
         false
     }